Apply visibility, embargo, and lease when creating transactionally
This duplicates the core behavior of `InterpretVisibilityActor`, allowing users
of transactional creation to pass in appropriate visibility arguments to achieve
embargo, lease, or just custom visibility.

Existing behavior, where simple visibility can be set on a work before a
transaction begins, remains.

Several details of the `InterpretVisibilityActor` remain unimplemented. For
instance, we don't check visibility/embargo/lease against a
`PermissionTemplate`. It's not immediately clear whether these features should
be re-added in the transactional context. It appears they cause
silent/unrecoverable failures in the Actor Stack context; if so, the status quo
is that clients responsible for not running afoul of them.