Add Collection PermissionTemplate application to transaction create
In some cases, Collection permission templates should be applied to works. The
actor stack has a complex way of handling this, split across the
`CollectionsMembershipActor` and the `ApplyPermissionTemplateActor`, see
especially: `CollectionsMembershipActor#extract_collection_id`.

Rather than reproduce this logic, we expect callers to the `create_work`
transaction to simply tell us which collections we should apply a permission
template for. This allows different user input to specify how collection
membership should effect permissions, even allowing multiple templates to be
applied (which is not possible in the actor stack's approach).